requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
flask
flask-cors
python-dotenv
requests
google-analytics-data
python-docx
pdfplumber
openpyxl
beautifulsoup4
lxml
chardet
waitress
gunicorn

